pyproject.toml
[project]
name = "smartclinicagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "faiss-cpu==1.11.0.post1",
    "fastembed==0.7.1",
    "flake8>=7.3.0",
    "google-ai-generativelanguage==0.6.15",
    "isort>=6.0.1",
    "langchain-community==0.3.27",
    "langchain[google-genai]==0.3.27",
    "langgraph==0.6.4",
    "langgraph-cli[inmem]>=0.3.6",
    "loguru==0.7.3",
    "markdown>=3.8.2",
    "pydantic-settings==2.10.1",
    "pylint>=3.3.8",
    "python-dotenv>=1.1.1",
    "pytz>=2025.2",
    "unstructured==0.18.11",
]
